requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Bcrypt==1.0.1
WTForms==3.2.1
email-validator==2.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
weasyprint==65.1
pytest==8.3.5
pytest-flask==1.3.0
coverage==7.8.0
openai==1.68.2
pypdf==5.4.0
openpyxl==3.1.5
fpdf2==2.8.2
PyMuPDF==1.26.5
boto3==1.39.11
